Added corner cases where we don't have any users in the databank thus handling choices differently

+++ MenuClass.py
@@ -34,19 +34,22 @@ class EntryMenu(Menu):
         return super().set_choice()
         
     def handle_choice(self, userdatabank: UserDatabank):
+        users = userdatabank.return_users()
         match self.choice:
             case '1':
                 userdatabank.show_users()
-                users = userdatabank.return_users()
-                choice = int(input("Choose: "))
-                if 1 <= choice <= len(users):
-                    print(users[choice - 1])
+                if len(users) != 0:
+                    choice = int(input("Choose: "))
+                    if 1 <= choice <= len(users):
+                        print(users[choice - 1])
             case '2':
                 userdatabank.add_user()
             case '3':
-                userdatabank.delete_user()
+                if len(users) != 0:
+                    userdatabank.delete_user()
             case '4':
                 sys.exit("Exiting the program")
+        return None
 
 
 class UserMenu(Menu):
